requirements.txt
sounddevice
numpy
faster-whisper
pyttsx3
groq
python-dotenv
edge-tts
pygame
